Lecture 01, Home Assignment
Task 01: only one item can be selected at any moment, next click on the same item must deselect it.
To solve this I used a ternary operator that checks if an iterated item is the one that was selected.
If it is — inverse selected property. Otherwise set it to false.

Task 02: Ensure that every item has a unique code/key.
While this goal can be achieved in multiple ways (like adding some generator function or just getting last/max code in the state list), I think in this case a simpler way is to keep an internal counter that is tied to the length of an initState array and increment it on every add button press. So this counter can be assigned to the new items and it should be unique within this project.

Task 03: Show number of times an item was selected.
To complete this I added an extra check in the selectItem method to increase the selectedCount property of an item if it is currently selected. Considering that we never explicitly set selectedCount prop, additional check is in order to make sure that it is set to 0 on the first selection (I used Nullish Coalescing Assignment operator).
Additionally the new function checks the value of count before rendering to set word 'раз' in proper declension by using the power of regex.

Лекция 01, домашнее задание.  

Задача 01: может быть только один выбранный элемент, следующее нажатие снимает выделение.  
Для получения нужного функционала я добавил тернарный оператор, который определяет, что сделать со значением свойства selected. Для выбранной записи значение инвертируется, для всех остальных устанавливается в false.  

Задача 02: каждая запись должна иметь уникальный идентификатор/ключ в пределах приложения.  
Данную задачу можно решить несколькими путями (например добавить функцию, которая будет генерировать уникальную последовательность чисел или делать проверку последнего/максимального идентификатора), но в пределах нашего приложения, я считаю, что один из самых простых способов будет добавить отдельное свойство, которое изначально будет привязано к длине начального массива + 1. Каждый новый элемент получает значение этого свойства, после чего оно увеличивается для следующего элемента.  

Задача 03: для каждого элемента выводить сколько раз он был выделен.
В методе selectItem я добавил доп проверку. Если запись выделена, то значение её свойства selectedCount увеличивается на один. Так как мы нигде явно не указывали такое свойство, то тут необходимо установить это свойство на 0 при первом выделении.  Дополнительно я добавил функцию, которая ставить слово "раз" в нужное склонение.