Lecture 02, Home Assignment by Melnichenko A
Key notes.

Cart items are stored in a property of the store.state.
This is an array of unique items with additional property count to keep track of the amount of the same item in the cart, rather than to store each new item as a separate entry.
While it's not quite ideal to add extra prop like that (and logic inside can be optimized...), I think it is a bit better to do it like this to achieve required goals in rendering cart related data. And, by my estimation, it would take more code to implement the same logic inside the cart component.

Controls takes cart content and cart opening handler function. Cart summary rendering can be further decomposed into a separate component and passed as an optional child to the Controls Component. And by doing this I would achieve better reusability.

In order to reuse components as much as possible, I made some changes to List and Item components.
List now takes the title for the control button and appropriate handler function. In turn both are passed into an Item component where the control button is rendered. Additionally count property conditionally rendered when Item comp. is used in the Cart.

Cart.
I opted to use an HTML5 <dialog> element to render the Cart.
Also, I took the liberty to show not only the total sum but the total amount of items in the cart. So total values are aligned with list values. But if this is not allowed/needed, it is easy to remove total count without breaking the layout.

new util functions:
thousSeparator - this one was made by me a long time ago for another project (so I might need to refactor it at some point, it is a bit hard to read).

calcCartSum, calcCartTotalCount - calculate total sum and count in the cart.

Заметки.
Товары в корзине хранятся в отдельном свойстве состояния стора в виде массива уникальных записей и дополнительным полем count для учитывания количества добавленных одинаковых товаров.
Это, конечно, не совсем идеально добавлять доп. свойство таким образом, но, я считаю, для выполения поставленных задач так будет лучше, чем хранить отдельные записи и делать логику группировки товаров внутри компонента Cart.

Controls дополнительно получает состояние корзины для вывода данных о содержимом корзины. Так же можно вынести эту часть в отдельный компонент, которые будет опциональным child этого компонента. И таким образом улучшить переиспользуемость компонента.

Я внёс небольшие изменения в компнонеты Item и List, для того, чтобы их можно было использовать и в корзине, а не создавать новые. Теперь List получает название для кнопки и связанную функцию, эти параметры дальше передаются в Item, где уже выводиться необходимая кнопка. Так же count кондиционно отображается, если он существуют у переданного элемента.

Корзина.
Для отображания модалки корзины, я использовал элемент <dialog>.
Дополнительно я решил добавить отображение не только общей суммы, но общего количество товаров. Таким образом все столбцы выравнены и показывают логически подходящие данные. Если это лишнее, то можно убрать количество товаров и это не должно сломать вёрстку.

Дополнительные функции.
thousSeparator - визуально разделяет числа на тысячи. Писал давно для другого проекта, надо будет отрефаторить (на мой взгляд тяжело читать)...
calcCartSum, calcCartTotalCount - считают количество и общую сумму товаров в корзине.